fix: add cache tests for feature flags
