+++ tsconfig.json
@@ -17,7 +17,6 @@
     "incremental": true,
     "baseUrl": "",
     "types": ["node", "vitest/globals"],
-    "noUncheckedIndexedAccess": true,
     "paths": {
       "@api/*": ["src/api/*"],
       "@components/*": ["src/components/*"],
